+++ src/com/dezzy/skorp3/skorp3D/raycast/render/Camera.java
@@ -9,8 +9,8 @@ public class Camera {
 	public Vector2 plane;
 	private Vector2 sidedir;
 	
-	private double moveSpeed = 0.055;
-	private double rotSpeed = 0.005;
+	private float moveSpeed = 0.055f;
+	private float rotSpeed = 0.005f;
 	
 	public double fogDistance = 10;
 	
@@ -34,7 +34,7 @@ public class Camera {
 		return rotSpeed;
 	}
 	
-	public void setRotationSpeed(double speed) {
+	public void setRotationSpeed(float speed) {
 		rotSpeed = speed;
 	}
 	
@@ -42,15 +42,15 @@ public class Camera {
 		return moveSpeed;
 	}
 	
-	public void setMoveSpeed(double speed) {
+	public void setMoveSpeed(float speed) {
 		moveSpeed = speed;
 	}
 	
 	private void computeSideDir() {
-		double t = Math.PI/2.0;
+		float t = (float) (Math.PI/2.0);
 		
-		double x = dir.x*Math.cos(t) - dir.y*Math.sin(t);
-		double y = dir.x*Math.sin(t) + dir.y*Math.cos(t);
+		float x = (float) (dir.x*Math.cos(t) - dir.y*Math.sin(t));
+		float y = (float) (dir.x*Math.sin(t) + dir.y*Math.cos(t));
 		
 		sidedir = new Vector2(x,y);
 	}
@@ -60,15 +60,15 @@ public class Camera {
 	 * 
 	 * @param factor multiplied by the rotation speed, product used to rotate left
 	 */
-	public void rotateLeft(double factor) {
-		double speed = factor * this.rotSpeed;
+	public void rotateLeft(float factor) {
+		float speed = factor * this.rotSpeed;
 		
-		double oldDirX = dir.x;
-		double cr = Math.cos(speed);
-		double sr = Math.sin(speed);
+		float oldDirX = dir.x;
+		float cr = (float) Math.cos(speed);
+		float sr = (float) Math.sin(speed);
 		dir.x = dir.x * cr - dir.y * sr;
 	    dir.y = oldDirX * sr + dir.y * cr;
-	    double oldPlaneX = plane.x;
+	    float oldPlaneX = plane.x;
 	    plane.x = plane.x * cr - plane.y * sr;
 	    plane.y = oldPlaneX * sr + plane.y * cr;
 	}
@@ -78,21 +78,21 @@ public class Camera {
 	 * 
 	 * @param factor multiplied by the rotation speed, product used to rotate right
 	 */
-	public void rotateRight(double factor) {
-		double speed = factor * this.rotSpeed;
+	public void rotateRight(float factor) {
+		float speed = factor * this.rotSpeed;
 		
-		double oldDirX = dir.x;
-	    double cr = Math.cos(-speed);
-	    double sr = Math.sin(-speed);
+		float oldDirX = dir.x;
+		float cr = (float) Math.cos(-speed);
+		float sr = (float) Math.sin(-speed);
 	    dir.x = dir.x * cr - dir.y * sr;
 	    dir.y = oldDirX * sr + dir.y * cr;
-	    double oldPlaneX = plane.x;
+	    float oldPlaneX = plane.x;
 	    plane.x = plane.x * cr - plane.y * sr;
 	    plane.y = oldPlaneX * sr + plane.y * cr;
 	}
 	
-	public void moveForward(WorldMap map, double factor) {
-		double speed = factor * moveSpeed;
+	public void moveForward(WorldMap map, float factor) {
+		float speed = factor * moveSpeed;
 		
 		if (map.get((int)(pos.x + dir.x * speed), (int)pos.y).id() == 0) {
 			pos.x += dir.x * speed;
@@ -102,8 +102,8 @@ public class Camera {
 		}
 	}
 	
-	public void moveBackward(WorldMap map, double factor) {
-		double speed = factor * moveSpeed;
+	public void moveBackward(WorldMap map, float factor) {
+		float speed = factor * moveSpeed;
 		
 		if (map.get((int)(pos.x - dir.x * speed),(int)pos.y).id() == 0) {
 			pos.x -= dir.x * speed;
@@ -113,35 +113,35 @@ public class Camera {
 	    }
 	}
 	
-	public void moveLeft(WorldMap map, double factor) {
+	public void moveLeft(WorldMap map, float factor) {
 		
 	}
 	
-	public void moveForward(double factor) {
-		double speed = factor * moveSpeed;
+	public void moveForward(float factor) {
+		float speed = factor * moveSpeed;
 		
 		pos.x+=dir.x*speed;
 		pos.y+=dir.y*speed;
 	}
 	
-	public void moveBackward(double factor) {
-		double speed = factor * moveSpeed;
+	public void moveBackward(float factor) {
+		float speed = factor * moveSpeed;
 		
 		pos.x-=dir.x*speed;
 		pos.y-=dir.y*speed;
 	}
 	
-	public void moveLeft(double factor) {
+	public void moveLeft(float factor) {
 		computeSideDir();
-		double speed = factor * moveSpeed;
+		float speed = factor * moveSpeed;
 		
 		pos.x+=sidedir.x*speed;
 		pos.y+=sidedir.y*speed;
 	}
 	
-	public void moveRight(double factor) {
+	public void moveRight(float factor) {
 		computeSideDir();
-		double speed = factor * moveSpeed;
+		float speed = factor * moveSpeed;
 		
 		pos.x-=sidedir.x*speed;
 		pos.y-=sidedir.y*speed;
